Move code under Tinkerbell cmd:
