+++ .golangci.yml
@@ -156,7 +156,7 @@ issues:
     - smee/internal/iso/internal/reverseproxy_test.go
     - smee/internal/iso/internal/acsii.go
     - smee/internal/iso/internal/acsii_test.go
-    - cmd/flag/url.go
+    - cmd/tinkerbell/flag/url.go
 
   # Excluding configuration per-path, per-linter, per-text and per-source
   exclude-rules:
